Added all the resources for the light culling... Added the rest of the code for the shaders. Fixed a small detail in File.h (max read and write using win api is max size of uint32_t. Added Light class. Added a way to get lights from a world... right now it sucks but I can generate an array of lights and get it in the renderer. Remove the images loading in the TestScene that was taking a long time to load an slowing down startup a lot.
LightCulling still doesn't work perfectly but it's a first pass to debug and iterate on.

+++ Source/Core/File/File.cpp
@@ -11,7 +11,7 @@
 
 namespace Core
 {
-	bool File::Read(uint8* buffer, uint64 maxReadSize)
+	bool File::Read(uint8* buffer, uint32_t maxReadSize)
 	{
 		assert(IsValid() && IsReadMode());
 
@@ -24,7 +24,7 @@ namespace Core
 	//	return "";
 	//}
 
-	bool File::Write(const uint8* source, uint64 size)
+	bool File::Write(const uint8* source, uint32_t size)
 	{
 		assert(IsValid() && IsWriteMode());
 		uint32 bytesWritten = 0;
